src/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
pyodbc>=5.2,<6.0
azure-identity>=1.19,<2.0
